% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_qsscore.R
\name{group_qsscore}
\alias{group_qsscore}
\title{Calculate quality of latent variable sum scores by groups}
\usage{
group_qsscore(model)
}
\arguments{
\item{model}{A lavaan model}
}
\value{
A data frame
}
\description{
Calculate quality of latent variable sum scores by groups
}
