% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_qsscore.R
\name{q_sscore}
\alias{q_sscore}
\title{Calculate quality of a composite score}
\usage{
q_sscore(lambda, var_data, variance_cs)
}
\arguments{
\item{lambda}{Standardized estimates from a model}

\item{var_data}{Variance of the actual data}

\item{variance_cs}{Variance of the composite score}
}
\value{
A numeric vector of length 1
}
\description{
Calculate quality of a composite score
}
