% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lavaan.R
\name{summary_lavaan}
\alias{summary_lavaan}
\title{Dig through lavaan objects and extract information as data frames.}
\usage{
summary_lavaan(object, ...)
}
\arguments{
\item{object}{A model of class class \code{\link{lavaan-class}}.}

\item{...}{all arguments from `lavaan::summary`. All argument information
is available at ?`lavaan-class` in the `summary` section.}
}
\value{
a list containing data frames with relevant information.
}
\description{
`lavaan` related objects contain a lot of information, specially in
group-wise models and models containing a lot of customization. Users
usually use `lavaan::summary`` to print information form the model but then
find themselves rambling through the inners of the model to actually
extract the same estimates that were previewed in `lavaan::summary`
as actual numbers and data frames. `summary_lavaan` is a rewrite
of `lavaan::summary` but instead of printing, it returns the same
information as a list with the information as data frames.
}
\examples{
library(lavaan)

model <- "
   # latent variables
    ind60 =~ x1 + x2 + x3
     dem60 =~ y1 + y2 + y3 + y4
     dem65 =~ y5 + y6 + y7 + y8
   # regressions
     dem60 ~ ind60
     dem65 ~ ind60 + dem60
   # residual covariances
     y1 ~~ y5
     y2 ~~ y4 + y6
     y3 ~~ y7
     y4 ~~ y8
     y6 ~~ y8"


fit <- sem(model, data=PoliticalDemocracy)
# Traditional print
summary(fit)

# Same information but as a data frames
summary_lavaan(fit)

}
